requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo[srv]==3.12.0
web3==6.0.0
python-dotenv==0.19.2